use renamedFrom and refactor
